chore: change quantity type to USD
We should probably introduce a `Quantity` type if we want to support different contracts. For now, a contract equals $1, hence, USD is correct.

+++ mobile/lib/common/domain/model.dart
@@ -108,7 +108,7 @@ class Usd {
 
   String formatted() {
     final formatter = NumberFormat("#,###,###,###,###", "en");
-    return formatter.format(_usd);
+    return formatter.format(_usd.toDouble());
   }
 
   @override
